+++ CHANGELOG.md
@@ -1,11 +1,33 @@
 #  Changelog
 
-## [0.8] - 2025-03-01
+## [0.9.10] - 2025-03-20
+
+### Added
+
+- Add your profile picture
+- Edit profile is now working with profile pictures.
+- Edited tab bar
+- Bug fixes to comments
+- Edited ProfileView
+- Edited RegisterView
+- release candidate version
+
+
+## [0.9] - 2025-03-16
+
+### Added
+
+- Reactions on comments
+
+## [0.8] - 2025-03-06
 
 ### Added
 
 - Changelog
-
+- Gleap chat
+- Quip+ renamed to Quip Premium
+- Edited tab bar
+- Notifications 
 
 ## [0.7] - 2025-02-028